backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
sentence-transformers
chromadb
pydantic
python-multipart
pypdf
python-docx
ollama
python-jose
passlib[bcrypt]